requirements.txt
pydantic>=2.0
pyyaml
openai
python-dotenv
pytest
pandas
streamlit
plotly
numpy
fredapi          # optional: FRED macro data init (requires FRED_API_KEY env var)
